requirements.txt
django~=4.1.1
django_environ~=0.9.0
djongo~=1.3.6
dj-database-url~=1.0.0
google-api-python-client~=2.61.0
gunicorn~=20.1.0
psycopg2~=2.9.3
pymongo~=3.12.3
pymongo[srv]
pytube~=12.1.0
pytz~=2022.2.1
timeago~=1.0.16
whitenoise~=6.2.0
httpx==0.23.0
